fixed mpsc crash
